fix: use BigDecimal instead of double for price/quantity fields
double is binary floating-point and can't exactly represent decimal
fractions, which compounds into rounding errors across arithmetic and
comparisons — a well-known pitfall for money-related values.

Related to #3

+++ backend/src/main/java/com/sdp/common/PriceTick.java
@@ -1,6 +1,7 @@
 package com.sdp.common;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
-public record PriceTick(String symbol, double bid, double ask, Instant timestamp) {
+public record PriceTick(String symbol, BigDecimal bid, BigDecimal ask, Instant timestamp) {
 }
